Move runners to separate modules
